backend/requirements.txt
<<<<<<< HEAD
=======
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
>>>>>>> 2ec1cf972b40c8cc42b0d1c2fd7982358d09fd50
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
<<<<<<< HEAD
djangorestframework_simplejwt==5.5.1
=======
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
executing==2.2.1
fonttools==4.61.1
greenlet==3.3.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.10.1
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
patsy==1.0.2
pillow==12.1.1
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.1
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.16.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
sklearn-preprocessing==0.1.0
soupsieve==2.8.3
SQLAlchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
statsmodels==0.14.6
threadpoolctl==3.6.0
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.2.14
zeep==4.3.2
>>>>>>> 2ec1cf972b40c8cc42b0d1c2fd7982358d09fd50
